# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "StudyForge Pro 智能学习平台 REST API。支持 AI 分析、间隔重复、流式对话、知识图谱等功能。\n\n所有认证端点需在 Header 携带 `Authorization: Bearer \u003ctoken\u003e`。",
        "title": "StudyForge Pro API",
        "contact": {
            "name": "StudyForge Pro",
            "url": "https://github.com/studyforge-pro"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取知识卡片，支持按材料、难度、待复习状态、书签过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识卡片"
                ],
                "summary": "卡片列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按材料 ID 过滤",
                        "name": "material_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按难度过滤（easy/medium/hard）",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "true 时只返回待复习卡片",
                        "name": "due",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "true 时只返回已书签卡片",
                        "name": "bookmarked",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "卡片列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cards/{id}/review": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "提交复习结果，基于 SM-2 算法更新间隔天数和难度因子",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "知识卡片"
                ],
                "summary": "复习卡片（SM-2）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "卡片 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "复习结果",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ReviewCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "复习结果",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "卡片不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/chat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发送消息并获取 AI 回复，支持 Function Calling 工具调用和 RAG 检索增强",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 对话"
                ],
                "summary": "对话（非流式）",
                "parameters": [
                    {
                        "description": "对话请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AI 回复",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "验证用户名密码，返回 JWT Token 和用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取当前用户的学习材料列表，支持标签过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "材料列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20，上限 200）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量（默认 0）",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按标签过滤",
                        "name": "tag",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "材料列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一条文本或 URL 类型的学习材料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "创建文本材料",
                "parameters": [
                    {
                        "description": "上传请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UploadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/model.Material"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/materials/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "上传 PDF/DOCX/MD/TXT 文件（最大 20MB），自动提取文本内容",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "上传文件材料",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "材料标题",
                        "name": "title",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "上传成功",
                        "schema": {
                            "$ref": "#/definitions/model.Material"
                        }
                    },
                    "400": {
                        "description": "文件无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/materials/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取单个材料详情，含关联的知识卡片和练习题",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "材料详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "材料 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "材料详情",
                        "schema": {
                            "$ref": "#/definitions/model.Material"
                        }
                    },
                    "404": {
                        "description": "材料不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除材料及其关联的卡片、练习题和 RAG 向量数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "删除材料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "材料 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "材料不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/materials/{id}/analyze": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "触发 4 个 Agent 并发分析材料（分析师/出题师/卡片师/图谱师），通过 WebSocket 推送进度",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "触发 AI 分析",
                "parameters": [
                    {
                        "type": "string",
                        "description": "材料 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "分析已启动",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "材料不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/quizzes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取练习题，支持按材料和难度过滤，可选智能推荐模式",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "练习题"
                ],
                "summary": "题目列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页条数（默认 20）",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按材料 ID 过滤",
                        "name": "material_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "按难度过滤",
                        "name": "difficulty",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "true 时启用智能推荐模式",
                        "name": "recommended",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "题目列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/quizzes/{id}/answer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "提交练习题答案，自动判断正确性并记录答题记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "练习题"
                ],
                "summary": "提交答案",
                "parameters": [
                    {
                        "type": "string",
                        "description": "题目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户答案",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "答题结果",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "题目不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的学习统计数据（材料数、卡片数、答题数、待复习卡片等）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "学习统计"
                ],
                "summary": "学习统计",
                "responses": {
                    "200": {
                        "description": "统计数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "聚合当前用户所有材料的标签及使用次数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "材料管理"
                ],
                "summary": "获取标签列表",
                "responses": {
                    "200": {
                        "description": "标签列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ChatRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "conversation_id": {
                    "description": "可选：指定对话会话（不传则自动创建）",
                    "type": "string"
                },
                "material_id": {
                    "description": "可选：关联材料上下文",
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "maxLength": 10000
                }
            }
        },
        "handler.ReviewCardRequest": {
            "type": "object",
            "required": [
                "result"
            ],
            "properties": {
                "result": {
                    "description": "\"mastered\" 或 \"review\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer Token（格式: Bearer xxx）",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}